Extract format_loves and fix type error
